7.17 更新

+++ resources/views/index/business/salesman-index.blade.php
@@ -86,11 +86,14 @@
                                                    data-id="{{ $man->id }}">
                                                     <i class="iconfont icon-chaopiao"></i>续费</a>
                                             @endif
+                                            @if(auth()->user()->type != cons('user.type.maker'))
+                                                <a data-url="{{ url('api/v1/business/salesman/'. $man->id) }}"
+                                                   data-method="delete"
+                                                   class="red ajax" type="button"><i class="iconfont icon-shanchu"></i>
+                                                    删除
+                                                </a>
+                                            @endif
 
-                                            <a data-url="{{ url('api/v1/business/salesman/'. $man->id) }}"
-                                               data-method="delete"
-                                               class="red ajax" type="button"><i class="iconfont icon-shanchu"></i> 删除
-                                            </a>
                                         </div>
                                     </td>
                                 </tr>
